+++ pybricksdev/connections/pybricks.py
@@ -36,6 +36,7 @@ from pybricksdev.ble.pybricks import (
     Event,
     HubCapabilityFlag,
     StatusFlag,
+    UserProgramId,
     unpack_hub_capabilities,
     unpack_pnp_id,
 )
@@ -471,18 +472,29 @@ class PybricksHub:
             response=True,
         )
 
-    async def start_user_program(self) -> None:
+    async def start_user_program(self, slot: UserProgramId | int | None = None) -> None:
         """
         Starts the user program that is already in RAM on the hub.
 
+        Args:
+            slot: The user program slot to start. If None, the currently selected slot is used.
+
         Requires hub with Pybricks Profile >= v1.2.0.
+
+        User program ID requires hub with Pybricks Profile >= v1.4.0.
         """
+
+        if slot is None:
+            slot = self._selected_slot
+
+        # TODO: add fallback for starting REPL on older firmware
+
         await self.write_gatt_char(
             PYBRICKS_COMMAND_EVENT_UUID,
             (
                 struct.pack("<B", Command.START_USER_PROGRAM)
                 if self._num_of_slots == 0
-                else struct.pack("<BB", Command.START_USER_PROGRAM, self._selected_slot)
+                else struct.pack("<BB", Command.START_USER_PROGRAM, slot)
             ),
             response=True,
         )